put test::executable_test into integration test to avoid parallel execution issue
